requirements.txt
boto3==1.9.217
botocore==1.12.217
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-anymail==6.1.0
django-dbbackup==3.2.0
docutils==0.15.2
idna==2.8
jmespath==0.9.4
Pillow==6.1.0
python-dateutil==2.8.0
pytz==2019.2
reportlab==3.5.23
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
